version information in the rootpage
